+++ CHANGELOG.asciidoc
@@ -26,6 +26,7 @@ NEED AND IMAGE
 This release also includes changes from <<release-3-3-2, 3.3.2>>.
 
 * Change the `toString()` of `Path` to be standardized as other graph elements are.
+* `hadoop-gremlin` no longer generates a test artifact.
 * Fixed a bug in `ReducingBarrierStep`, that returned the provided seed value despite no elements being available.
 * Changed the order of `select()` scopes. The order is now: maps, side-effects, paths.
 * Removed support for Giraph.
