+++ gremlin-test/src/main/java/org/apache/tinkerpop/gremlin/structure/io/gryo/ToyTriangle.java
@@ -17,21 +17,17 @@
  *  under the License.
  */
 
-package org.apache.tinkerpop.gremlin.hadoop.structure.io.gryo;
+package org.apache.tinkerpop.gremlin.structure.io.gryo;
 
-import org.apache.tinkerpop.gremlin.structure.Vertex;
 import org.apache.tinkerpop.gremlin.structure.io.graphson.AbstractObjectDeserializer;
-import org.apache.tinkerpop.gremlin.structure.io.graphson.GraphSONTokens;
 import org.apache.tinkerpop.gremlin.structure.io.gryo.kryoshim.InputShim;
 import org.apache.tinkerpop.gremlin.structure.io.gryo.kryoshim.KryoShim;
 import org.apache.tinkerpop.gremlin.structure.io.gryo.kryoshim.OutputShim;
 import org.apache.tinkerpop.gremlin.structure.io.gryo.kryoshim.SerializerShim;
-import org.apache.tinkerpop.gremlin.structure.util.detached.DetachedVertex;
 import org.apache.tinkerpop.shaded.jackson.core.JsonGenerationException;
 import org.apache.tinkerpop.shaded.jackson.core.JsonGenerator;
 import org.apache.tinkerpop.shaded.jackson.databind.SerializerProvider;
 import org.apache.tinkerpop.shaded.jackson.databind.ser.std.StdScalarSerializer;
-import org.apache.tinkerpop.shaded.jackson.databind.ser.std.StdSerializer;
 
 import java.io.IOException;
 import java.util.Map;
@@ -39,14 +35,16 @@ import java.util.Map;
 /**
  * @author Marko A. Rodriguez (http://markorodriguez.com)
  */
-public final class ToyPoint {
+public final class ToyTriangle {
 
     private final int x;
     private final int y;
+    private final int z;
 
-    public ToyPoint(final int x, final int y) {
+    public ToyTriangle(final int x, final int y, final int z) {
         this.x = x;
         this.y = y;
+        this.z = z;
     }
 
     public int getX() {
@@ -57,57 +55,65 @@ public final class ToyPoint {
         return this.y;
     }
 
+    public int getZ() {
+        return this.z;
+    }
+
     public int hashCode() {
-        return this.x + this.y;
+        return this.x + this.y + this.z;
     }
 
     public boolean equals(final Object other) {
-        return other instanceof ToyPoint && ((ToyPoint) other).x == this.x && ((ToyPoint) other).y == this.y;
+        return other instanceof ToyTriangle && ((ToyTriangle) other).x == this.x && ((ToyTriangle) other).y == this.y && ((ToyTriangle) other).z == this.z;
     }
 
     @Override
     public String toString() {
-        return "[" + this.x + "," + this.y + "]";
+        return "[" + this.x + "," + this.y + "," + this.z + "]";
     }
 
-    public static class ToyPointSerializer implements SerializerShim<ToyPoint> {
+    public static class ToyTriangleSerializer implements SerializerShim<ToyTriangle> {
         @Override
-        public <O extends OutputShim> void write(final KryoShim<?, O> kryo, final O output, final ToyPoint toyPoint) {
-            output.writeInt(toyPoint.x);
-            output.writeInt(toyPoint.y);
+        public <O extends OutputShim> void write(final KryoShim<?, O> kryo, final O output, final ToyTriangle toyTriangle) {
+            output.writeInt(toyTriangle.x);
+            output.writeInt(toyTriangle.y);
+            output.writeInt(toyTriangle.z);
         }
 
         @Override
-        public <I extends InputShim> ToyPoint read(final KryoShim<I, ?> kryo, final I input, final Class<ToyPoint> toyPointClass) {
-            return new ToyPoint(input.readInt(), input.readInt());
+        public <I extends InputShim> ToyTriangle read(final KryoShim<I, ?> kryo, final I input, final Class<ToyTriangle> toyTriangleClass) {
+            return new ToyTriangle(input.readInt(), input.readInt(), input.readInt());
         }
     }
 
-    public static class ToyPointJacksonSerializer extends StdScalarSerializer<ToyPoint> {
 
-        public ToyPointJacksonSerializer() {
-            super(ToyPoint.class);
+    public static class ToyTriangleJacksonSerializer extends StdScalarSerializer<ToyTriangle> {
+
+        public ToyTriangleJacksonSerializer() {
+            super(ToyTriangle.class);
         }
 
         @Override
-        public void serialize(final ToyPoint toyPoint, final JsonGenerator jsonGenerator,
+        public void serialize(final ToyTriangle toyTriangle, final JsonGenerator jsonGenerator,
                               final SerializerProvider serializerProvider) throws IOException, JsonGenerationException {
             jsonGenerator.writeStartObject();
-            jsonGenerator.writeObjectField("x", toyPoint.x);
-            jsonGenerator.writeObjectField("y", toyPoint.y);
+            jsonGenerator.writeObjectField("x", toyTriangle.x);
+            jsonGenerator.writeObjectField("y", toyTriangle.y);
+            jsonGenerator.writeObjectField("z", toyTriangle.z);
             jsonGenerator.writeEndObject();
         }
     }
 
-    public static class ToyPointJacksonDeSerializer extends AbstractObjectDeserializer<ToyPoint> {
+    public static class ToyTriangleJacksonDeSerializer extends AbstractObjectDeserializer<ToyTriangle> {
 
-        public ToyPointJacksonDeSerializer() {
-            super(ToyPoint.class);
+        public ToyTriangleJacksonDeSerializer() {
+            super(ToyTriangle.class);
         }
 
         @Override
-        public ToyPoint createObject(final Map<String, Object> map) {
-            return new ToyPoint((int) map.get("x"), (int) map.get("y"));
+        public ToyTriangle createObject(final Map<String, Object> map) {
+            return new ToyTriangle((int) map.get("x"), (int) map.get("y"), (int) map.get("z"));
         }
     }
+
 }
